Added Shopping Cart functionalities

+++ src/main/java/com/example/webshopbackend/repository/UserRepository.java
@@ -11,5 +11,5 @@ public interface UserRepository extends JpaRepository <User, Long>{
     Optional<User> findUserByUsernameAndPassword(String username, String password);
 
     Optional<User> findUserByUsername(String username);
-    User findByUserId(int userId);
+    User findByUserId(Long userId);
 }
